+++ js/plugins/code/options-component.js
@@ -1250,7 +1250,10 @@ class OptionsComponent extends LitElement {
 
                         <div class="menu-item-static content-section">
                             <label>Update Wisk PWA</label>
-                            <button class="btn btn-developer" @click="${() => window.updateWiskPWA()}">Update</button>
+                            <button class="btn btn-developer" @click="${async () => {
+                                await window.clearWiskPWA();
+                                window.updateWiskPWA();
+                            }}">Update</button>
                         </div>
 
                         <div class="menu-item-static content-section">